LayoutTests/fast/events/event-function-toString-expected.txt

Test serialization of event handlers for bug 25544.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL div.onclick should be function onclick(event) { test(); } (of type string). Was function onclick(event) { test();
} (of type function).
FAIL path.onclick should be function onclick(evt) { doIt(); } (of type string). Was function onclick(evt) { doIt();
} (of type function).
PASS successfullyParsed is true

TEST COMPLETE
